+++ packages/stream_feeds/lib/src/state/feed_state.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:clock/clock.dart';
 import 'package:collection/collection.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:state_notifier/state_notifier.dart';
@@ -671,7 +672,7 @@ extension on FeedState {
     final updatedNotificationStatus = notificationStatus?.copyWith(
       unread: 0,
       readActivities: readActivities,
-      lastReadAt: DateTime.timestamp(),
+      lastReadAt: clock.now().toUtc(),
     );
 
     return copyWith(notificationStatus: updatedNotificationStatus);
@@ -691,7 +692,7 @@ extension on FeedState {
     final updatedNotificationStatus = notificationStatus?.copyWith(
       unseen: 0,
       seenActivities: seenActivities,
-      lastSeenAt: DateTime.timestamp(),
+      lastSeenAt: clock.now().toUtc(),
     );
 
     return copyWith(notificationStatus: updatedNotificationStatus);
@@ -715,7 +716,7 @@ extension on FeedState {
     final updatedNotificationStatus = notificationStatus?.copyWith(
       unread: updatedUnreadCount,
       readActivities: updatedReadActivities,
-      lastReadAt: DateTime.timestamp(),
+      lastReadAt: clock.now().toUtc(),
     );
 
     return copyWith(notificationStatus: updatedNotificationStatus);
@@ -739,7 +740,7 @@ extension on FeedState {
     final updatedNotificationStatus = notificationStatus?.copyWith(
       unseen: updatedUnseenCount,
       seenActivities: updatedSeenActivities,
-      lastSeenAt: DateTime.timestamp(),
+      lastSeenAt: clock.now().toUtc(),
     );
 
     return copyWith(notificationStatus: updatedNotificationStatus);